+++ Regula_Falsi/util.py
@@ -61,7 +61,7 @@ class util:
             returns: float
         """
         return abs(
-            Decimal(x_prev) - Decimal(x)
+            float(x_prev) - float(x)
         )
 
     def verifyOppositeSign(self, a, b):
